design

+++ app/src/main/java/com/alef/souqleader/data/remote/dto/Language.kt
@@ -3,5 +3,5 @@ package com.alef.souqleader.data.remote.dto
 data class Language(
     val id: Int,
     val name: String,
-    val proficiency: Int
+    val proficiency: Float
 )
