+++ app/src/main/java/com/alef/souqleader/data/remote/dto/Skill.kt
@@ -3,5 +3,5 @@ package com.alef.souqleader.data.remote.dto
 data class Skill(
     val id: Int,
     val name: String,
-    val progress: Int
+    val progress: Float
 )
